Get rid of babeljs
