+++ .gitignore
@@ -1,3 +1,1 @@
-/lib/*
 node_modules/
-hubot.js
