--- .jsbeautifyrc
@@ -1,9 +1,0 @@
-{
-  "brace_style": "collapse-preserve-inline",
-  "end_with_newline": true,
-  "indent_char": " ",
-  "indent_size": 4,
-  "jslint_happy": false,
-  "space_after_anon_function": false,
-  "space_before_conditional": true
-}
