+++ .npmignore
@@ -1,2 +1,1 @@
-src
 test
