+++ lib/jira.js
@@ -1,6 +1,6 @@
-import JiraApi from 'jira-client';
+const JiraApi = require('jira-client');
 
-export default function(config) {
+module.exports = function(config) {
     const jira = new JiraApi({
         protocol: 'https',
         host: config.jiraHost,
@@ -11,12 +11,10 @@ export default function(config) {
     });
 
     function getIssuesByKeys(keys) {
-        return jira
-            .searchJira(`key in (${keys.join(',')})`)
-            .then(res => res.issues);
+        return jira.searchJira(`key in (${keys.join(',')})`).then(res => res.issues);
     }
 
     return {
-        getIssuesByKeys: keys => getIssuesByKeys(keys),
+        getIssuesByKeys: keys => getIssuesByKeys(keys)
     };
-}
+};
