+++ lib/messageFormatter.js
@@ -1,8 +1,8 @@
-import _ from 'lodash';
+const _ = require('lodash');
 
 let config;
 
-export default (configuration) => {
+module.exports = configuration => {
     config = configuration;
 
     return {
@@ -21,10 +21,12 @@ function buildDetailedMessage(issue) {
         title: `${issue.key}: ${issue.fields.summary}`,
         title_link: `https://${config.jiraHost}/browse/${issue.key}`,
         text: issue.fields.description,
-        fields: [{
-            title: 'Status',
-            value: issue.fields.status.name,
-        }]
+        fields: [
+            {
+                title: 'Status',
+                value: issue.fields.status.name
+            }
+        ]
     };
 }
 
@@ -34,7 +36,9 @@ function buildMessage(issue) {
         color: getColorForStatus(issue.fields.status.name),
         author_name: issue.fields.assignee ? issue.fields.assignee.displayName : 'Unassigned',
         author_icon: issue.fields.assignee ? issue.fields.assignee.avatarUrls['16x16'] : null,
-        text: `<https://${config.jiraHost}/browse/${issue.key}|${issue.key}> \`${issue.fields.status.name}\` ${issue.fields.summary}`,
+        text: `<https://${config.jiraHost}/browse/${issue.key}|${issue.key}> \`${issue.fields.status.name}\` ${
+            issue.fields.summary
+        }`,
         mrkdwn_in: ['text']
     };
 }
@@ -59,8 +63,7 @@ function getColorForStatus(status) {
 }
 
 function isUnknownKeyError(errorMessage) {
-    return errorMessage &&
-        (errorMessage.indexOf('does not exist') !== -1 || errorMessage.indexOf('is invalid') !== -1);
+    return errorMessage && (errorMessage.indexOf('does not exist') !== -1 || errorMessage.indexOf('is invalid') !== -1);
 }
 
 function buildErrorMessage(error) {
@@ -68,7 +71,7 @@ function buildErrorMessage(error) {
         errorMessage = _.get(error, 'errorMessages[0]');
 
     if (isUnknownKeyError(errorMessage)) {
-        const chunks = errorMessage.split('\'');
+        const chunks = errorMessage.split("'");
         message = `:exclamation: \`${chunks[1]}\` cannot be found`;
     }
